requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pandas==2.2.2
gunicorn==22.0.0


# Testing & Tooling
pytest==8.3.3
requests==2.32.3

